+++ R/tidypredict.R
@@ -12,6 +12,8 @@
 #' @importFrom stats qt
 #' @importFrom knitr knit_print
 #' @importFrom dplyr mutate
+#' @importFrom dplyr case_when
+#' @importFrom partykit nodeapply
 #' @keywords internal
 #'
 "_PACKAGE"